refactor: move painting info to widgets layer
