+++ src/widgets/fields/index.ts
@@ -1,2 +1,3 @@
 export { default as GameField } from "./ui/GameField";
 export { default as WordsPicker } from "./ui/WordsPicker";
+export { default as PaintingInfo } from "./ui/PaintingInfo";
