+++ src/widgets/fields/ui/PaintingInfo.module.css
